# apps/core-ledger-api/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Core Ledger API",
    "description": "Financial Ledger REST API for institutional financial clients",
    "contact": {
      "name": "Core Ledger Team",
      "email": "support@coreledger.com"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://localhost:7109",
      "description": "Local development (HTTPS)"
    },
    {
      "url": "http://localhost:5071",
      "description": "Local development (HTTP)"
    }
  ],
  "paths": {
...
    "/api/v1/calendario": {
      "get": {
        "tags": [
          "Calendário"
        ],
        "operationId": "GetAllCalendarios",
        "parameters": [
          {
            "name": "Limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "SortBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortDirection",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Praca",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "TipoDia",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "DiaUtil",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "DataInicio",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "DataFim",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoreLedger.Application.Models.PagedResult`1[[CoreLedger.Application.DTOs.CalendarioDto, CoreLedger.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Calendário"
        ],
        "operationId": "CreateCalendario",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoreLedger.Application.DTOs.CreateCalendarioDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoreLedger.Application.DTOs.CalendarioDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/calendario/{id}": {
      "get": {
        "tags": [
          "Calendário"
        ],
        "operationId": "GetCalendarioById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoreLedger.Application.DTOs.CalendarioDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Calendário"
        ],
        "operationId": "UpdateCalendario",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoreLedger.Application.DTOs.UpdateCalendarioDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Calendário"
        ],
        "operationId": "DeleteCalendario",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/calendario/dia-util/{data}": {
      "get": {
        "tags": [
          "Calendário"
        ],
        "operationId": "CheckDiaUtil",
        "parameters": [
          {
            "name": "data",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "praca",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/CoreLedger.Domain.Enums.Praca"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/calendario/proximo-dia-util/{data}": {
      "get": {
        "tags": [
          "Calendário"
        ],
        "operationId": "GetProximoDiaUtil",
        "parameters": [
          {
            "name": "data",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "praca",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/CoreLedger.Domain.Enums.Praca"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/calendario/calcular-d-mais/{data}/{dias}": {
      "get": {
        "tags": [
          "Calendário"
        ],
        "operationId": "CalcularDMais",
        "parameters": [
          {
            "name": "data",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dias",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "praca",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/CoreLedger.Domain.Enums.Praca"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/calendario/importar/{ano}": {
      "post": {
        "tags": [
          "Calendário"
        ],
        "operationId": "ImportarCalendario",
        "parameters": [
          {
            "name": "ano",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "praca",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/CoreLedger.Domain.Enums.Praca"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CoreLedger.Application.DTOs.CalendarioDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "string",
            "format": "date"
          },
          "diaUtil": {
            "type": "boolean"
          },
          "tipoDia": {
            "$ref": "#/components/schemas/CoreLedger.Domain.Enums.TipoDia"
          },
          "tipoDiaDescricao": {
            "type": "string",
            "nullable": true
          },
          "praca": {
            "$ref": "#/components/schemas/CoreLedger.Domain.Enums.Praca"
          },
          "pracaDescricao": {
            "type": "string",
            "nullable": true
          },
          "descricao": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CoreLedger.Application.DTOs.CreateCalendarioDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "string",
            "format": "date"
          },
          "tipoDia": {
            "$ref": "#/components/schemas/CoreLedger.Domain.Enums.TipoDia"
          },
          "praca": {
            "$ref": "#/components/schemas/CoreLedger.Domain.Enums.Praca"
          },
          "descricao": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CoreLedger.Application.DTOs.UpdateCalendarioDto": {
        "type": "object",
        "properties": {
          "tipoDia": {
            "$ref": "#/components/schemas/CoreLedger.Domain.Enums.TipoDia"
          },
          "descricao": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CoreLedger.Domain.Enums.Praca": {
        "enum": [
          1,
          2,
          3,
          4,
          5
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "CoreLedger.Domain.Enums.TipoDia": {
        "enum": [
          1,
          2,
          3,
          4,
          5,
          6,
          7
        ],
        "type": "integer",
        "format": "int32"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme. Enter 'Bearer' [space] and then your token in the text input below.\n\nExample: \"Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\"",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}